+++ src/variable_versions/v9.rs
@@ -19,7 +19,6 @@ use nom::multi::many0;
 use nom_derive::*;
 use serde::Serialize;
 
-use std::collections::BTreeMap;
 use std::collections::HashMap;
 
 pub const DATA_TEMPLATE_V9_ID: u16 = 0;
@@ -323,16 +322,13 @@ impl<'a> OptionsFieldParser {
     fn parse(
         input: &'a [u8],
         template: &OptionsTemplate,
-    ) -> IResult<&'a [u8], Vec<BTreeMap<usize, V9FieldPair>>> {
+    ) -> IResult<&'a [u8], Vec<Vec<V9FieldPair>>> {
         let mut result = Vec::new();
         let mut remaining = input;
-        for (count, template_field) in template.option_fields.iter().enumerate() {
+        for template_field in template.option_fields.iter() {
             let (i, field_value) = template_field.parse_as_field_value(remaining)?;
             remaining = i;
-            result.push(BTreeMap::from([(
-                count,
-                (template_field.field_type, field_value),
-            )]));
+            result.push(vec![(template_field.field_type, field_value)]);
         }
         Ok((remaining, result))
     }
@@ -346,7 +342,7 @@ pub struct OptionsDataFields {
     pub scope_fields: Vec<ScopeDataField>,
     // Options Data Fields
     #[nom(Parse = "{ |i| OptionsFieldParser::parse(i, template) }")]
-    pub options_fields: Vec<BTreeMap<usize, V9FieldPair>>,
+    pub options_fields: Vec<Vec<V9FieldPair>>,
 }
 
 #[derive(Debug, PartialEq, Clone, Serialize)]
@@ -480,7 +476,7 @@ impl<'a> FieldParser {
     ///
     /// The function computes the number of records available in the input by dividing the length of the input
     /// by the template's total size. It then iteratively extracts each record using `parse_data_field`, accumulating
-    /// a vector of BTreeMaps where each map represents a record mapping field indices to `V9FieldPair`.
+    /// a vector of where each map represents a record mapping field indices to `V9FieldPair`.
     ///
     /// # Arguments
     ///
@@ -491,7 +487,7 @@ impl<'a> FieldParser {
     ///
     /// A result containing:
     /// - The remaining slice of input data that was not parsed.
-    /// - A vector of BTreeMaps, each mapping a field index to its corresponding `V9FieldPair`.
+    /// - A vector of V9FieldPair, each mapping a field index to its corresponding `V9FieldPair`.
     ///
     /// # Errors
     ///
@@ -522,8 +518,7 @@ impl<'a> FieldParser {
     /// Parses a single record (data field) based on the provided template.
     ///
     /// The function iterates over each field defined in the template, using each field's own parser to
-    /// extract its value from the input. The parsed values, along with their corresponding field types,
-    /// are accumulated in a BTreeMap keyed by field index.
+    /// extract its value from the input. The parsed values, along with their corresponding field types.
     ///
     /// # Arguments
     ///
@@ -534,7 +529,7 @@ impl<'a> FieldParser {
     ///
     /// A result containing:
     /// - The remaining input slice after parsing the record.
-    /// - A BTreeMap mapping field indices to `V9FieldPair`, where each pair consists of the field type and the parsed value.
+    /// - A Vector of `V9FieldPair`, where each pair consists of the field type and the parsed value.
     ///
     /// # Errors
     ///
@@ -636,8 +631,7 @@ impl V9 {
                         }
                     }
                     for options_field in options_data_field.options_fields.iter() {
-                        for (index, (_field_type, field_value)) in options_field.iter() {
-                            result.extend_from_slice(&index.to_be_bytes());
+                        for (_field_type, field_value) in options_field.iter() {
                             result.extend_from_slice(&field_value.to_be_bytes()?);
                         }
                     }